+++ crates/grammars/src/diff/outline.scm
@@ -1,0 +1,7 @@
+(block
+    (command (_)) @name
+) @item
+
+(hunk
+    (location (_)) @name
+) @item